Move buildAccount code out to prepare family split
