+++ src/libcore/reconciliation.js
@@ -1,0 +1,81 @@
+// @flow
+// libcore reconciliation by the React definition. https://reactjs.org/docs/reconciliation.html
+
+import isEqual from "lodash/isEqual";
+import type { Operation } from "../types";
+
+const sameImmutableOp = (a: Operation, b: Operation) =>
+  a === b ||
+  (a.id === b.id && // hash, accountId, type are in id
+    a.date.getTime() === b.date.getTime() &&
+    (a.fee ? a.fee.isEqualTo(b.fee) : a.fee === b.fee) &&
+    (a.value ? a.value.isEqualTo(b.value) : a.value === b.value) &&
+    isEqual(a.senders, b.senders) &&
+    isEqual(a.recipients, b.recipients));
+
+function findExistingOp(ops, op) {
+  return ops.find(o => o.id === op.id);
+}
+
+export async function minimalOperationsBuilder<CO>(
+  existingOperations: Operation[],
+  coreOperations: CO[],
+  buildOp: (coreOperation: CO) => Promise<Operation>
+): Promise<Operation[]> {
+  // build operations with the minimal diff & call to libcore possible
+  let operations = [];
+  let existingOps = existingOperations || [];
+
+  let immutableOpCmpDoneOnce = false;
+  for (let i = coreOperations.length - 1; i >= 0; i--) {
+    const coreOperation = coreOperations[i];
+    const newOp = await buildOp(coreOperation);
+    const existingOp = findExistingOp(existingOps, newOp);
+
+    if (existingOp && !immutableOpCmpDoneOnce) {
+      // an Operation is supposely immutable.
+      if (existingOp.blockHeight !== newOp.blockHeight) {
+        // except for blockHeight that can temporarily be null
+        operations.push(newOp);
+        continue; // eslint-disable-line no-continue
+      } else {
+        immutableOpCmpDoneOnce = true;
+        // we still check the first existing op we meet...
+        if (!sameImmutableOp(existingOp, newOp)) {
+          // this implement a failsafe in case an op changes (when we fix bugs)
+          // tradeoff: in such case, we assume all existingOps are to trash
+          console.warn("op mismatch. doing a full clear cache.");
+          existingOps = [];
+          operations.push(newOp);
+          continue; // eslint-disable-line no-continue
+        }
+      }
+    }
+
+    if (existingOp) {
+      // as soon as we've found a first matching op in old op list,
+      const j = existingOps.indexOf(existingOp);
+      const rest = existingOps.slice(j);
+      if (rest.length !== i + 1) {
+        // if libcore happen to have different number of ops that what we have,
+        // we actualy need to continue because we don't know where hole will be,
+        // but we can keep existingOp
+        operations.push(existingOp);
+      } else {
+        // otherwise we stop the libcore iteration and continue with previous data
+        // and we're done on the iteration
+        if (operations.length === 0 && j === 0) {
+          // special case: we preserve the operations array as much as possible
+          operations = existingOps;
+        } else {
+          operations = operations.concat(rest);
+        }
+        break;
+      }
+    } else {
+      // otherwise it's a new op
+      operations.push(newOp);
+    }
+  }
+  return operations;
+}
